+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Advection of a collection of vortex blobs: `advection!`. ([#5])
 
 - Concrete type for a Cartesian mesh: `CartesianMesh`. ([#9])
+- Abstract type representing a vortex-blob redistribution kernel: `AbstractRedistributionKernel`. ([#9])
+- Concrete type for the M4' redistribution kernel: `M4Prime`. ([#9])
 
 ### Changed